backend/requirements.txt
fastapi
python-dotenv
uvicorn
python-multipart
clarifai
SQLAlchemy
PyJWT
passlib
pydantic[email]
bcrypt==4.0.1
resend
jinja2
openai
